Zadanie10/pyproject.toml
[tool.poetry]
name = "pwzn-pb-z10"
version = "0.1.0"
description = "Python w zastosowaniach naukowych - Zadanie 10"
authors = ["Piotr Bienias"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
beautifulsoup4 = "^4.12.3"
requests = "^2.31.0"
pillow = "^10.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
